:construction: Begin cleanup of pre-approval submission dialog.
